# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UnifyOS Crisis Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus];

export const DeviceStatus = {
  online: "online",
  offline: "offline",
  warning: "warning",
  critical: "critical",
} as const;

export interface Device {
  id: string;
  name: string;
  location: string;
  status: DeviceStatus;
  lastSeen: string;
  createdAt: string;
}

export interface RegisterDeviceRequest {
  id: string;
  name: string;
  location: string;
}

export interface SensorDataRequest {
  /** Temperature in Celsius */
  temperature: number;
  /** Smoke level in PPM */
  smoke: number;
  /** Motion detected (0 or 1) */
  motion: number;
  /** Panic button pressed (0 or 1) */
  button: number;
}

export interface SensorReading {
  id: number;
  deviceId: string;
  temperature: number;
  smoke: number;
  motion: number;
  button: number;
  timestamp: string;
}

export type AnomalyResultSeverity =
  (typeof AnomalyResultSeverity)[keyof typeof AnomalyResultSeverity];

export const AnomalyResultSeverity = {
  NORMAL: "NORMAL",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  CRITICAL: "CRITICAL",
} as const;

export interface AnomalyResult {
  severity: AnomalyResultSeverity;
  confidence: number;
  anomalies: string[];
  action: string;
  message: string;
}

export interface SensorDataResponse {
  reading: SensorReading;
  anomaly: AnomalyResult;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  NORMAL: "NORMAL",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  CRITICAL: "CRITICAL",
} as const;

export interface Alert {
  id: number;
  deviceId: string;
  deviceName: string;
  deviceLocation: string;
  severity: AlertSeverity;
  confidence: number;
  anomalies: string[];
  message: string;
  action: string;
  dismissed: boolean;
  createdAt: string;
}

export type GetDeviceReadingsParams = {
  limit?: number;
};

export type ListAlertsParams = {
  limit?: number;
  deviceId?: string;
};