# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UnifyOS Crisis Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: deviceId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{alertId}/dismiss:
    post:
      operationId: dismissAlert
      tags: [alerts]
      summary: Dismiss an alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
components:
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: string
        deviceName:
          type: string
        deviceLocation:
          type: string
        severity:
          type: string
          enum: [NORMAL, MEDIUM, HIGH, CRITICAL]
        confidence:
          type: number
        anomalies:
          type: array
          items:
            type: string
        message:
          type: string
        action:
          type: string
        dismissed:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - deviceName
        - deviceLocation
        - severity
        - confidence
        - anomalies
        - message
        - action
        - dismissed
        - createdAt