# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UnifyOS Crisis Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from "./alertSeverity";

export interface Alert {
  id: number;
  deviceId: string;
  deviceName: string;
  deviceLocation: string;
  severity: AlertSeverity;
  confidence: number;
  anomalies: string[];
  message: string;
  action: string;
  dismissed: boolean;
  createdAt: Date;
}